time: fix ParseDuration overflow when given more than 9 digits on 32-bit arch Fixes #6617.
R=golang-dev, rsc, r
CC=golang-dev
https://golang.org/cl/15080043

+++ src/pkg/time/format.go
@@ -1204,11 +1204,11 @@ func ParseDuration(s string) (Duration, error) {
 			if err != nil {
 				return 0, errors.New("time: invalid duration " + orig)
 			}
-			scale := 1
+			scale := 1.0
 			for n := pl - len(s); n > 0; n-- {
 				scale *= 10
 			}
-			g += float64(x) / float64(scale)
+			g += float64(x) / scale
 			post = pl != len(s)
 		}
 		if !pre && !post {
